Exclude binary cache build/install for live ebuilds
This commit prevents binary cache from being built for sources that are
fetched from VCS (for eg., git). It is no longer necessary to pass a
list of packages with live ebuilds to --buildpkg-exclude or
--usepkg-exclude.

Before this commit, when an emerge is called with the option
'--buildpkg', a binary cache for the package is created under
/var/cache/binpkgs. For example, when we do a, 'emerge --ask
--verbose --buildpkg some-gitpkg/abc', a binary cache abc-1.1.1.tbz2
is created under /var/cache/binpkgs/some-gitpkg.

And, if we do a 'emerge --ask --verbose --usepkg some-gitpkg/abc',
the package is then installed from the binary that is available already.

With this commit, even if we explicitly use the options, '--buildpkg'
and '--usepkg' for the packages with live ebuilds, no binary cache
will be created or installed. It will always be fetched from the source.
We no longer have to pass a list of VCS based packages to be excluded
via '--buildpkg-exclude' and '--usepkg-exclude'.

Motivation: Since binary caches are created for all packages, including
packages with live ebuilds, a separate list of (for eg.,) git packages
needs to be maintained. And this is then passed to the options
'--buildpkg-exclude' and '--usepkg-exclude' via, EMERGE_DEFAULT_OPTS.
So the motivation behind this patch was to reduce redundancy, while we
can simply disable binary cache for sources fetched from a VCS.

Also, if a stale binary cache for git based package is present in the
system, it possible that we might end up using this old cache with
'--usepkg', instead of rebuilding it from sources.

Signed-off-by: Madhu Priya Murugan <madhu.murugan@rohde-schwarz.com>